app/requirements.txt
Flask
pytest
python-magic